Project_Gift_Price/requirements.txt
python-telegram-bot>=20.0
pillow>=9.0.0
numpy>=1.22.0
requests>=2.28.0
python-dotenv>=0.20.0
matplotlib>=3.5.0
httpx>=0.23.0 